Fixed makefiles and API changes for compatibility with latest devkitpro

+++ SDL-1.2.15/src/audio/switch/SDL_switchaudio.h
@@ -41,7 +41,7 @@ struct SDL_PrivateAudioData {
     Uint8  bytePerSample;
 	AudioOutBuffer waveBuf[NUM_BUFFERS];
 	AudioOutBuffer *released_buffer;
-	u32 released_count;
+	u64 released_count;
 	u32 next;
 };
